sort the data before serialization, then merge (numeric)
This reworks the state_numeric - instead of (Numeric*) array with
pointers to the Numeric values, we keep just a buffer with copies
of the data. We need the buffers, and we need to repalloc() them
in _append_ methods, which however invalidates the pointers. And
we only need the pointers for sorting, so instead of maintaining
them all the time, we simply build them ad-hoc.