when both state and value are NULL, _append_ can return NULL directly
This is a minor optimization - we don't allocate state until we get
the first non-NULL element. So at the beginning of the append function
we simply do

    if (PG_ARGISNULL(0) && PG_ARGISNULL(1))
        PG_RETURN_NULL();

This check was already there for NUMERIC, this commit also adds it
to the other data types (INT, BIGINT, DOUBLE PRECISION), and changes
the comment a bit.