handle negative numerators in STDDEV/VAR final functions
It can happen that due to round-off errors, we get

    N * sum(X*X) - sum(X)*sum(X) < 0

In those cases we must not use this to compute the VAR/STDDEV
estimates, and instead return 0.

For the NUMERIC data type, we use a global static variable
const_zero, so that we don't have to recompute it over and over.