add sort_state_* functions wrapping the pg_qsort calls
For double, int32 and int64 this is fairly straightforward, but
for Numeric it requires reordering data in the buffer (in-place).

This also allows removal of build_numeric_elements() function,
and instead read the data directly from the buffer. The code is
a bit more complicated as we need to skip some elements, though.