fix 'value overflows numeric format' error in numeric aggregates
The numeric aggregates with numerator (N*sumX2 - sumX*sumX) were
failing with this strange error

    ERROR:  value overflows numeric format

when evaluating the numerator. Not sure what exactly is the cause,
but getting rid of the numerator fixed it.

Also, get rid of the const_zero idea and simply return new value
every time we hit the round-off condition (which should not be
very common). People may do pfree() on the result, which would
immediately break the const_zero value.